Fix regression in win32 build caused by string class changes. Now using ptrdiff_t, which automatically selects correct signed int type.

+++ ceng-win32/src/gl32-pixel-shader.cpp
@@ -46,7 +46,7 @@ const Ceng::CRESULT GL32_PixelShader::GetInstance(const Ceng::StringUtf8 &shader
 
 	size_t length = shaderText.Length();
 
-	if (length > std::numeric_limits<GLint>::max())
+	if (length > size_t(std::numeric_limits<GLint>::max()))
 	{
 		return CE_ERR_INVALID_PARAM;
 	}
